+++ src/main/java/com/zepsun/aberrance/block/custom/CustomPlacementBushBlock.java
@@ -10,32 +10,19 @@ import net.minecraft.world.level.block.state.BlockState;
 import java.util.List;
 
 public class CustomPlacementBushBlock extends BushBlock {
-    public static List<Object> mayPlaceList;
+    public static TagKey<Block> mayPlaceTag;
 
-    public CustomPlacementBushBlock(List<Object> pMayPlaceList, Properties pProperties) {
+    public CustomPlacementBushBlock(TagKey<Block> pMayPlaceTag, Properties pProperties) {
         super(pProperties);
-        mayPlaceList = pMayPlaceList;
+        mayPlaceTag = pMayPlaceTag;
     }
 
     @Override
     protected boolean mayPlaceOn(BlockState pState, BlockGetter pLevel, BlockPos pPos) {
-        int i;
-        for (i = 0; i < mayPlaceList.size(); i++)
-        {
-            if (mayPlaceList.get(i) instanceof Block) {
-                if (pState.is((Block) mayPlaceList.get(i))) {
-                    return true;
-                }
-            } else if (mayPlaceList.get(i) instanceof TagKey<?>) {
-                if (pState.is((TagKey<Block>) mayPlaceList.get(i))) {
-                    return true;
-                }
-            }
-        }
-        return false;
+        return pState.is(mayPlaceTag);
     }
 
-    public static List<Object> getMayPlaceList() {
-        return mayPlaceList;
+    public static TagKey<Block> getPlacementTag() {
+        return mayPlaceTag;
     }
 }