+++ src/main/java/com/zepsun/aberrance/datagen/DataGenerators.java
@@ -35,6 +35,8 @@ public class DataGenerators {
         generator.addProvider(event.includeClient(), new ModItemModelProvider(packOutput, existingFileHelper));
         generator.addProvider(event.includeClient(), new ModBlockStateProvider(packOutput, existingFileHelper));
 
+        generator.addProvider(event.includeClient(), new ModEnUsGenerator(packOutput));
+
         //generator.addProvider(event.includeServer(), new ModWorldGenProvider(packOutput, lookupProvider));
 
     }