+++ src/main/java/com/zepsun/aberrance/datagen/ModEnUsGenerator.java
@@ -1,0 +1,162 @@
+package com.zepsun.aberrance.datagen;
+
+import com.zepsun.aberrance.Aberrance;
+import com.zepsun.aberrance.block.ModBlocks;
+import com.zepsun.aberrance.item.ModItems;
+import net.minecraft.data.PackOutput;
+import net.minecraftforge.common.data.LanguageProvider;
+
+public class ModEnUsGenerator extends LanguageProvider {
+
+    public ModEnUsGenerator(PackOutput output) {
+        super(output, Aberrance.MOD_ID, "en_us");
+    }
+
+    @Override
+    protected void addTranslations() {
+
+        addBlock(ModBlocks.PARSNIP_CROP, "Parsnip Crop");
+        addBlock(ModBlocks.WILD_PARSNIP_PLANT, "Wild Parsnip Plant");
+        addItem(ModItems.PARSNIP, "Parsnip");
+        addItem(ModItems.PARSNIP_SEEDS, "Parsnip Seeds");
+
+        addBlock(ModBlocks.WINTER_WHEAT_CROP, "Winter Wheat Crop");
+        addBlock(ModBlocks.WILD_WINTER_WHEAT_PLANT, "Wild Winter Wheat Plant");
+        addItem(ModItems.WINTER_WHEAT, "Winter Wheat");
+        addItem(ModItems.WINTER_WHEAT_BREAD, "Winter Wheat Bread");
+        addItem(ModItems.WINTER_WHEAT_SEEDS, "Parsnip Seeds");
+
+        addBlock(ModBlocks.ONION_CROP, "Onion Crop");
+        addBlock(ModBlocks.WILD_ONION_PLANT, "Wild Onion Plant");
+        addItem(ModItems.ONION, "Onion");
+
+        addBlock(ModBlocks.ACACIA_MESQUITE_LEAVES, "Acacia Mesquite Leaves");
+        addBlock(ModBlocks.ACACIA_MESQUITE_SAPLING, "Acacia Mesquite Sapling");
+        addItem(ModItems.MESQUITE, "Mesquite");
+
+        addBlock(ModBlocks.SEED_MAKER, "Seed Maker");
+        addBlock(ModBlocks.COBBLESTONE_CASE, "Cobblestone Case");
+
+        addBlock(ModBlocks.KYANITE_ORE, "Kyanite Ore");
+        addBlock(ModBlocks.KYANITE_ORE_DEEPSLATE, "Deepslate Kyanite Ore");
+        addBlock(ModBlocks.KYANITE_ORE_NETHER, "Netherrack Kyanite Ore");
+        addBlock(ModBlocks.KYANITE_ORE_END, "End Stone Kyanite Ore");
+        addBlock(ModBlocks.RAW_KYANITE_BLOCK, "Raw Kyanite Block");
+        addBlock(ModBlocks.KYANITE_BLOCK, "Kyanite Block");
+        addItem(ModItems.KYANITE_INGOT, "Kyanite Ingot");
+        addItem(ModItems.KYANITE_NUGGET, "Kyanite Nugget");
+        addItem(ModItems.RAW_KYANITE, "Raw Kyanite");
+        addItem(ModItems.KYANITE_PICKAXE, "Kyanite Sword");
+        addItem(ModItems.KYANITE_SWORD, "Kyanite Pickaxe");
+        addItem(ModItems.KYANITE_SHOVEL, "Kyanite Shovel");
+        addItem(ModItems.KYANITE_AXE, "Kyanite Axe");
+        addItem(ModItems.KYANITE_HOE, "Kyanite Hoe");
+        addItem(ModItems.KYANITE_HELMET, "Kyanite Helmet");
+        addItem(ModItems.KYANITE_CHESTPLATE, "Kyanite Chestplate");
+        addItem(ModItems.KYANITE_LEGGINGS, "Kyanite Leggings");
+        addItem(ModItems.KYANITE_BOOTS, "Kyanite Boots");
+
+        addBlock(ModBlocks.RUDANIUM_ORE, "Rudanium Ore");
+        addBlock(ModBlocks.RUDANIUM_ORE_DEEPSLATE, "Deepslate Rudanium Ore");
+        addBlock(ModBlocks.RUDANIUM_ORE_NETHER, "Netherrack Rudanium Ore");
+        addBlock(ModBlocks.RUDANIUM_ORE_END, "End Stone Rudanium Ore");
+        addBlock(ModBlocks.RAW_RUDANIUM_BLOCK, "Raw Rudanium Block");
+        addBlock(ModBlocks.RUDANIUM_BLOCK, "Rudanium Block");
+        addItem(ModItems.RUDANIUM_INGOT, "Rudanium Ingot");
+        addItem(ModItems.RUDANIUM_NUGGET, "Rudanium Nugget");
+        addItem(ModItems.RAW_RUDANIUM, "Raw Rudanium");
+        addItem(ModItems.RUDANIUM_PICKAXE, "Rudanium Sword");
+        addItem(ModItems.RUDANIUM_SWORD, "Rudanium Pickaxe");
+        addItem(ModItems.RUDANIUM_SHOVEL, "Rudanium Shovel");
+        addItem(ModItems.RUDANIUM_AXE, "Rudanium Axe");
+        addItem(ModItems.RUDANIUM_HOE, "Rudanium Hoe");
+        addItem(ModItems.RUDANIUM_HELMET, "Rudanium Helmet");
+        addItem(ModItems.RUDANIUM_CHESTPLATE, "Rudanium Chestplate");
+        addItem(ModItems.RUDANIUM_LEGGINGS, "Rudanium Leggings");
+        addItem(ModItems.RUDANIUM_BOOTS, "Rudanium Boots");
+
+        addBlock(ModBlocks.CREPE_MYRTLE_LOG, "Crepe Myrtle Log");
+        addBlock(ModBlocks.CREPE_MYRTLE_WOOD, "Crepe Myrtle Wood");
+        addBlock(ModBlocks.STRIPPED_CREPE_MYRTLE_LOG, "Stripped Crepe Myrtle Log");
+        addBlock(ModBlocks.STRIPPED_CREPE_MYRTLE_WOOD, "Stripped Crepe Myrtle Wood");
+        addBlock(ModBlocks.CREPE_MYRTLE_SAPLING, "Crepe Myrtle Sapling");
+        addBlock(ModBlocks.CREPE_MYRTLE_PLANKS, "Crepe Myrtle Planks");
+        addBlock(ModBlocks.CREPE_MYRTLE_STAIRS, "Crepe Myrtle Stairs");
+        addBlock(ModBlocks.CREPE_MYRTLE_SLAB, "Crepe Myrtle Slab");
+        addBlock(ModBlocks.CREPE_MYRTLE_PRESSURE_PLATE, "Crepe Myrtle Pressure Plate");
+        addBlock(ModBlocks.CREPE_MYRTLE_BUTTON, "Crepe Myrtle Button");
+        addBlock(ModBlocks.CREPE_MYRTLE_DOOR, "Crepe Myrtle Door");
+        addBlock(ModBlocks.CREPE_MYRTLE_TRAPDOOR, "Crepe Myrtle Trapdoor");
+        addBlock(ModBlocks.CREPE_MYRTLE_FENCE, "Crepe Myrtle Fence");
+        addBlock(ModBlocks.CREPE_MYRTLE_FENCE_GATE, "Crepe Myrtle Fence Gate");
+        addBlock(ModBlocks.CREPE_MYRTLE_LEAVES, "Crepe Myrtle Leaves");
+        addBlock(ModBlocks.CREPE_MYRTLE_COMPOST, "Crepe Myrtle Compost");
+
+        addBlock(ModBlocks.LIVE_OAK_LOG, "Live Oak Log");
+        addBlock(ModBlocks.LIVE_OAK_WOOD, "Live Oak Wood");
+        addBlock(ModBlocks.STRIPPED_LIVE_OAK_LOG, "Stripped Live Oak Log");
+        addBlock(ModBlocks.STRIPPED_LIVE_OAK_WOOD, "Stripped Live Oak Wood");
+        addBlock(ModBlocks.LIVE_OAK_SAPLING, "Live Oak Sapling");
+        addBlock(ModBlocks.LIVE_OAK_PLANKS, "Live Oak Planks");
+        addBlock(ModBlocks.LIVE_OAK_STAIRS, "Live Oak Stairs");
+        addBlock(ModBlocks.LIVE_OAK_SLAB, "Live Oak Slab");
+        addBlock(ModBlocks.LIVE_OAK_PRESSURE_PLATE, "Live Oak Pressure Plate");
+        addBlock(ModBlocks.LIVE_OAK_BUTTON, "Live Oak Button");
+        addBlock(ModBlocks.LIVE_OAK_DOOR, "Live Oak Door");
+        addBlock(ModBlocks.LIVE_OAK_TRAPDOOR, "Live Oak Trapdoor");
+        addBlock(ModBlocks.LIVE_OAK_FENCE, "Live Oak Fence");
+        addBlock(ModBlocks.LIVE_OAK_FENCE_GATE, "Live Oak Fence Gate");
+        addBlock(ModBlocks.LIVE_OAK_LEAVES, "Live Oak Leaves");
+
+        addBlock(ModBlocks.BROADLEAF_LOG, "Broadleaf Log");
+        addBlock(ModBlocks.BROADLEAF_WOOD, "Broadleaf Wood");
+        addBlock(ModBlocks.STRIPPED_BROADLEAF_LOG, "Stripped Broadleaf Log");
+        addBlock(ModBlocks.STRIPPED_BROADLEAF_WOOD, "Stripped Broadleaf Wood");
+        addBlock(ModBlocks.BROADLEAF_SAPLING, "Broadleaf Sapling");
+        addBlock(ModBlocks.BROADLEAF_PLANKS, "Broadleaf Planks");
+        addBlock(ModBlocks.BROADLEAF_STAIRS, "Broadleaf Stairs");
+        addBlock(ModBlocks.BROADLEAF_SLAB, "Broadleaf Slab");
+        addBlock(ModBlocks.BROADLEAF_PRESSURE_PLATE, "Broadleaf Pressure Plate");
+        addBlock(ModBlocks.BROADLEAF_BUTTON, "Broadleaf Button");
+        addBlock(ModBlocks.BROADLEAF_DOOR, "Broadleaf Door");
+        addBlock(ModBlocks.BROADLEAF_TRAPDOOR, "Broadleaf Trapdoor");
+        addBlock(ModBlocks.BROADLEAF_FENCE, "Broadleaf Fence");
+        addBlock(ModBlocks.BROADLEAF_FENCE_GATE, "Broadleaf Fence Gate");
+        addBlock(ModBlocks.BROADLEAF_LEAVES, "Broadleaf Leaves");
+
+        addBlock(ModBlocks.WHITE_COBBLESTONE, "White Cobblestone");
+        addBlock(ModBlocks.WHITE_COBBLESTONE_WALL, "White Cobbl stone Wall");
+        addBlock(ModBlocks.WHITE_COBBLESTONE_SLAB, "White Cobblestone Slab");
+        addBlock(ModBlocks.WHITE_COBBLESTONE_STAIRS, "White Cobblestone Stairs");
+        addBlock(ModBlocks.WHITE_STONE, "White Stone");
+        addBlock(ModBlocks.WHITE_STONE_WALL, "White Stone Wall");
+        addBlock(ModBlocks.WHITE_STONE_SLAB, "White Stone Slab");
+        addBlock(ModBlocks.WHITE_STONE_STAIRS, "White Stone Stairs");
+        addBlock(ModBlocks.WHITE_STONE_BRICKS, "White Stone Bricks");
+        addBlock(ModBlocks.WHITE_STONE_BRICK_WALL, "White Stone Brick Wall");
+        addBlock(ModBlocks.WHITE_STONE_BRICK_SLAB, "White Stone Brick Slab");
+        addBlock(ModBlocks.WHITE_STONE_BRICK_STAIRS, "White Stone Brick Stairs");
+
+        addBlock(ModBlocks.SOUL_SEDIMENT, "Soul Sediment");
+
+        addBlock(ModBlocks.STARVED_SOUL_SOIL, "Starved Soul Soil");
+        addBlock(ModBlocks.STARVED_SOUL_GRASS, "Starved Soul Grass");
+        addBlock(ModBlocks.STARVED_SOUL_GRASS_BLOCK, "Starved Soul Grass Block");
+
+        addBlock(ModBlocks.MURDEROUS_SOUL_SOIL, "Murderous Soul Soil");
+        addBlock(ModBlocks.MURDEROUS_SOUL_GRASS, "Murderous Soul Grass");
+
+        addItem(ModItems.CUTLASS, "Cutlass");
+        addItem(ModItems.VIOLIN_BOW, "Violin Bow");
+
+        addItem(ModItems.MUSIC_DISC_ENRICHED, "Music Disc");
+        add("item.aberrance.music_disc_enriched.desc", "zepsun - Waltz of the Enriched");
+
+        add("aberrance.tooltip.amount.with.capacity", "%s / %s mB");
+        add("aberrance.tooltip.amount", "%s mB");
+        add("container.white_stone_furnace", "White Stone Furnace");
+        add("creativetab.aberrance", "Aberrance");
+        add("creativetab.aberrance_weapons", "Aberrance Weapons");
+
+    }
+}