+++ src/main/java/com/zepsun/aberrance/datagen/loot/ModBlockLootTables.java
@@ -154,6 +154,7 @@ public class ModBlockLootTables extends BlockLootSubProvider {
         this.dropSelf(ModBlocks.SOUL_SEDIMENT.get());
         this.dropSelf(ModBlocks.MURDEROUS_SOUL_GRASS.get());
         this.dropSelf(ModBlocks.STARVED_SOUL_GRASS.get());
+        this.dropSelf(ModBlocks.STARVED_SOUL_GRASS_BLOCK.get());