+++ src/main/java/com/zepsun/aberrance/util/PortalHandlers.java
@@ -1,13 +1,4 @@
 package com.zepsun.aberrance.util;
 
-import com.zepsun.aberrance.dimensiongen.ModDimensions;
-import net.minecraft.core.BlockPos;
-import net.minecraft.resources.ResourceKey;
-import net.minecraft.server.MinecraftServer;
-import net.minecraft.server.level.ServerLevel;
-import net.minecraft.world.level.Level;
-
-import javax.swing.text.html.parser.Entity;
-
 public class PortalHandlers {
 }